+++ src/HubActivity.kt
@@ -64,6 +64,10 @@ class HubActivity : MainListActivity("hub", R.id.hub) {
                 startActivity(Intent(this, ConversationsActivity::class.java).addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT))
                 true
             }
+            R.id.ic705_rx_diagnostic -> {
+                startActivity(Intent(this, org.aprsdroid.app.ic705.diagnostic.Ic705RxDiagnosticActivity::class.java))
+                true
+            }
             R.id.export -> {
                 onStartLoading()
                 LogExporter(this, StorageDatabase.open(this), null) { onStopLoading() }.execute()